create wrapper in model.py to change the categorical output of repo's mlp to binary output clean genre folder
